test(chore): start of redoing dataset copy tests
